// front/utils/types/coworkingFormErrorInterface.ts
interface ICoworkingsErrorInfo {
    name?: string;
    lastname?: string;
    phone?: string;
    email?: string;
    identification?: string;
    position?: string;
    companyName?: string;
    companyEmail?: string;
    companyPhone?: string;
    address?: string;
    website?: string;
    open?: string;
    close?: string;
    capacity?: string;
    status?: string;
    observation?: string;
    message?: string;
    type?: string;
  }
  
  export default ICoworkingsErrorInfo;